LPS-140232 deal with NPE

+++ modules/apps/object/object-service/src/main/java/com/liferay/object/internal/model/listener/UserModelListener.java
@@ -39,6 +39,10 @@ public class UserModelListener extends BaseModelListener<User> {
 				_objectDefinitionLocalService.fetchObjectDefinitionByClassName(
 					user.getCompanyId(), User.class.getName());
 
+			if (objectDefinition == null) {
+				return;
+			}
+
 			_objectEntryLocalService.deleteRelatedObjectEntries(
 				0, objectDefinition.getObjectDefinitionId(),
 				user.getPrimaryKey());
